+++ lang/nl/cms.php
@@ -5,19 +5,25 @@ declare(strict_types=1);
 return [
     'updated' => 'Site-informatie bijgewerkt.',
     'manage'  => [
+        'add_field'              => 'Site-informatieveld toevoegen',
         'actions'                => 'Acties',
-        'delete_field_warning'   => '<b>:field</b> verwijderen? De ingevulde waarde gaat verloren.',
-        'delete_subject_warning' => '<b>:subject</b> verwijderen? Alle onderliggende onderwerpen, velden, ingevulde waarden en gegenereerde SVG-bestanden binnen dit onderwerp gaan verloren.',
+        'delete_field_warning'   => ':field verwijderen? De ingevulde waarde gaat verloren.',
+        'delete_subject_warning' => ':subject verwijderen? Alle onderliggende onderwerpen, velden, ingevulde waarden en gegenereerde SVG-bestanden binnen dit onderwerp gaan verloren.',
         'title'                  => 'Site-informatie beheren',
     ],
     'fields' => [
+        'add'     => 'Veld toevoegen',
         'created' => 'Site-informatieveld aangemaakt.',
         'deleted' => 'Site-informatieveld verwijderd.',
+        'title'   => 'Velden voor :subject',
         'updated' => 'Site-informatieveld bijgewerkt.',
     ],
     'subjects' => [
-        'created' => 'Site-informatieonderwerp aangemaakt.',
-        'deleted' => 'Site-informatieonderwerp verwijderd.',
-        'updated' => 'Site-informatieonderwerp bijgewerkt.',
+        'add_child'      => 'Onderwerp toevoegen',
+        'back_to_subject' => 'Terug naar onderwerp',
+        'children_title' => 'Onderwerpen voor :subject',
+        'created'        => 'Site-informatieonderwerp aangemaakt.',
+        'deleted'        => 'Site-informatieonderwerp verwijderd.',
+        'updated'        => 'Site-informatieonderwerp bijgewerkt.',
     ],
 ];