+++ resources/views/cms/layouts/site-information/show.blade.php
@@ -13,6 +13,8 @@
                         class="btn btn-secondary"
                         href="{{ route('site-information.manage') }}"
                         role="button"
+                        title="{{ __('velor-site-information::cms.manage.title') }}"
+                        aria-label="{{ __('velor-site-information::cms.manage.title') }}"
                     >
                         <i class="bi bi-gear"></i>
                     </a>
@@ -22,6 +24,8 @@
                     class="btn btn-primary"
                     href="{{ route('site-information.edit') }}"
                     role="button"
+                    title="{{ __('cms.button_titles.edit', ['resource' => __('velor-site-information::resources.site-information-subjects.plural')]) }}"
+                    aria-label="{{ __('cms.button_titles.edit', ['resource' => __('velor-site-information::resources.site-information-subjects.plural')]) }}"
                 >
                     <i class="bi bi-pencil"></i>
                 </a>