+++ resources/views/cms/layouts/site-information/subject-show.blade.php
@@ -1,0 +1,36 @@
+@extends('app')
+
+@section('content')
+    <div class="titlebar sticky-top position-sticky">
+        <div class="d-flex align-items-center justify-content-between">
+            <div class="d-flex align-items-center">
+                <h3>{{ $show->getTitle() }}</h3>
+            </div>
+
+            <x-cms.layout.language-switcher />
+        </div>
+    </div>
+
+    <x-cms.layout.tabs :tabs="$show->getTabs()" />
+
+    <div class="content overflow-auto">
+        @foreach ($show->getPanels() as $panelIndex => $panel)
+            @include('components.cms.layout.show-panel', [
+                'panel' => $panel,
+                'panelIdPrefix' => 'show-panel-' . $panelIndex,
+            ])
+        @endforeach
+
+        <div class="d-flex flex-wrap gap-2">
+            @foreach ($show->getButtons() as $button)
+                <x-cms.layout.button :buttonData="$button" />
+            @endforeach
+
+            @can('delete', $model)
+                <x-cms.layout.button :buttonData="$deleteButton" />
+            @endcan
+        </div>
+    </div>
+
+    <x-cms.modal.delete />
+@endsection